# Improving performance for hash prefix validation.