@stebet 2021-05-27
Just noticed PR #9 which this PR would invalidate.